Shift logic to proper event
